backend/python-api-service/requirements.txt
# Core Flask dependencies
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Flask-SocketIO==5.3.6
python-socketio==5.8.0
python-engineio==4.7.1

# Database
psycopg2-binary==2.9.7
SQLAlchemy==2.0.23

# Authentication & Security
Werkzeug==2.3.7
bcrypt==4.0.1
PyJWT==2.8.0

# External API integrations
requests==2.31.0
httpx==0.25.2

# Data processing
pandas==2.1.4
numpy==1.25.2

# Email
Flask-Mail==0.9.1

# Caching & Background tasks
redis==5.0.1
celery==5.3.4

# File handling
python-multipart==0.0.6
Pillow==10.1.0

# Environment & Configuration
python-dotenv==1.0.0

# Development & Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-mock==3.12.0
coverage==7.3.2

# Code quality
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Documentation
sphinx==7.2.6

# Monitoring & Logging
sentry-sdk[flask]==1.38.0

# System monitoring
psutil==5.9.6

# Date & Time
python-dateutil==2.8.2

# URL parsing
urllib3==2.1.0

# JSON Schema validation
jsonschema==4.19.2

# Web scraping (for integrations)
beautifulsoup4==4.12.2
lxml==4.9.3

# PDF processing
PyPDF2==3.0.1

# Excel processing
openpyxl==3.1.2

# Image processing
opencv-python==4.8.1.78

# Voice processing
speechrecognition==3.10.0
pyttsx3==2.90

# Machine Learning (for agents)
scikit-learn==1.3.2
transformers==4.35.2
torch==2.1.1

# Natural Language Processing
nltk==3.8.1
spacy==3.7.2

# Calendar integrations
google-api-python-client==2.108.0
google-auth-oauthlib==1.2.0

# Communication integrations
slack-sdk==3.26.0
twilio==8.11.0

# Finance integrations
plaid-python==16.0.0

# Development tools
jupyter==1.0.0
ipykernel==6.26.0

# Deployment
gunicorn==21.2.0
gevent==23.9.1

# Cloud storage
boto3==1.34.11

# GraphQL (optional)
graphene==3.3
graphene-sqlalchemy==3.0.0rc1

# Rate limiting
flask-limiter==3.5.0

# API documentation
flask-restx==1.2.0

# Task scheduling
APScheduler==3.10.4

# Data validation
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0

# Caching
flask-caching==2.1.0

# Compression
flask-compress==1.14

# Health checks
flask-healthz==0.0.1

# Metrics
prometheus-flask-exporter==0.22.4
